Added back the old Health Cross
This version of the health indicator is disabled by default. To enable it, open flawhud/resource/ui/hudplayerhealth.res and look for the the HealthBG node. Disable it by setting its enabled and visible values to 0, then go to the PlayerStatusHealthBonusImage node. Under this node, find the image value and change it from "" to "../hud/health_over_bg" then restart the game.