Fixed #238 (Hopefully)
Fixed black bars appearing for item containers that display a "model" instead of an image of the item (namely skinned weapons). This seems to only happen in 1440p resolution and is caused by the bounding box of the item panel not sizing up properly with the space allocated to the item. Most of the tinkering was done with model_wide and model_tall to find that sweet spot.